refactor(LottoReference): rename result to reference
