+++ src/test/java/lotto/constants/LottoReferenceTest.java
@@ -2,12 +2,13 @@ package lotto.constants;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import lotto.model.LottoReference;
 import org.junit.jupiter.api.Test;
 
-class LottoResultTest {
+class LottoReferenceTest {
 
     @Test
     void getPrizeByCorrectCount_3_5000() {
-        assertThat((LottoResult.getPrizeByCorrectCount(3))).isEqualTo(5_000);
+        assertThat((LottoReference.getPrizeByCorrectCount(3))).isEqualTo(5_000);
     }
 }
